Moves testing related code to spec folder
