--- src/TestUtils.js
@@ -1,15 +1,0 @@
-import { destroyAllDataPermanently } from './DatabaseAdapter';
-
-let unsupported = function() {
-  throw 'Only supported in test environment';
-};
-
-let _destroyAllDataPermanently;
-if (process.env.TESTING) {
-  _destroyAllDataPermanently = destroyAllDataPermanently;
-} else {
-  _destroyAllDataPermanently = unsupported;
-}
-
-export default {
-    destroyAllDataPermanently: _destroyAllDataPermanently};
